fix: use with_utc_timestamps() with SimpleLogger
On some systems SimpleLogger can't be initialized and
complains with:

"Could not determine the UTC offset on this system. Possible
causes are that the time crate does not implement "local_offset_at"
on your system, or that you are running in a multi-threaded
environment and the time crate is returning "None" from
"local_offset_at" to avoid unsafe behaviour. See the time crate's
documentation for more information.
(https://time-rs.github.io/internal-api/time/index.html#feature-flags): IndeterminateOffset'"