add: fetch sanctioned addresses
Previously, the sanctioned addresses were stored in
`sanctioned_addresses_XBT.json` and used in code-generation during
compilation. They'd need to be updated manually and the binaries
would need to be recompiled.

A better approach might be to fetch the addresses from GitHub and
then storing them in the database. Updating them on restarts and
periodically, e.g., every 24h. This reduces the time between new
sanctioned addresses being added and miningpool-observer checking
for transactions to or from these addresses. It also reduces the
needed ongoing maintainance.

Addresses are currently fetched from
https://raw.githubusercontent.com/0xB10C/ofac-sanctioned-digital-currency-addresses/lists/sanctioned_addresses_XBT.txt
by default. This can be configured in the `daemon-config.toml`
file with the option `sanctioned_addresses_url`.