fix: regenerate schema.rs after upgrading to 2.0
During the upgrade to diesel 2.0 in the previous commits, the
schema.rs file wasn't regenerated. Diesel 2.0 now supports PostgreSQL
arrays containing NULL values. See https://diesel.rs/guides/migration_guide.html#changed-nullability-of-array-elements

We don't use NULL values in arrays and remove the Nullable wrapper in
a patch as described in the 1.4 to 2.0 migration guide. The patch might
need to be updated if more tables using PostgreSQL arrays are added.